Entries are now saved as a python dictionary.
This allows the user to re-enter case information if they accidentally mess up the first time. A new button "Submit All Entries" was added. This button writes the dictionary (the one with all the saved case information) into the CSV file. Every single case entry must be saved (even blank ones) in order to write to the CSV file.