Selected CSV file saves file information to dictionary.
Once a file is selected, the information in the file is saved to a dictionary. They user is then able to scan a slide and read the slides barcode. I still need to make it so the dictionary values are then saved in a class. Also need to make it so when a barcode is scanned it compares the scanned barcode to the barcodes saved in the class.