A scanned barcode is now compared with the barcodes in the setup file.
When the setup file is read, the barcode information is now saved directly into instance variables of the class 'Data'. When the user presses the 'read barcode' button after scanning it, the scanned barcode is compared with the barcodes saved in the class 'Data'. Depending on the barcode scanned, a message is displayed to the user to place the slide in the proper slot. The user can re-scan and read a new barcode as many times as they would like. Things I need to add: 1. illuminate the correct case when a barcode is scanned. 2. If a scanned barcode does not match the barcodes saved in the class 'Data', a message is portrayed to the user.